+++ src/db_walk/mod.rs
@@ -3,7 +3,6 @@ use colored::Colorize;
 use log::info;
 use std::path::PathBuf;
 use tokio::task::{self, JoinError, JoinSet};
-mod dag_custom;
 
 use crate::{
     config::SubsetConfig,
