--- v.txt
@@ -1,118 +1,0 @@
-digraph {
-    0 [ label = analyses_Crowns ]
-    1 [ label = public_PmsTreatments ]
-    2 [ label = public_DataUploadTriggers ]
-    3 [ label = public_ImagingPatients ]
-    4 [ label = NonPhi_ImageManifests ]
-    5 [ label = public_Practices ]
-    6 [ label = public_AnalyzedImages ]
-    7 [ label = public_ImageCatalog ]
-    8 [ label = analyses_Roots ]
-    9 [ label = public_PatientIdMatches ]
-    10 [ label = public_ImageAnalysisMirrors ]
-    11 [ label = public_EstimatedVisit ]
-    12 [ label = public_OrganizationPreferences ]
-    13 [ label = Phi_PhiImageMetadata ]
-    14 [ label = public_PmsProviders ]
-    15 [ label = analyses_Findings ]
-    16 [ label = public_ImagingPatientsSources ]
-    17 [ label = analyses_FmxSummary ]
-    18 [ label = analyses_Caries ]
-    19 [ label = public_PracticePreferences ]
-    20 [ label = public_PmsPatientViewCounts ]
-    21 [ label = public_QuarantineImages ]
-    22 [ label = public_MagicLink ]
-    23 [ label = analyses_Fillings ]
-    24 [ label = analyses_RootCanals ]
-    25 [ label = analyses_BoneLevels ]
-    26 [ label = public_UtilizationMetrics ]
-    27 [ label = public_DataUploadAudit ]
-    28 [ label = public_Organizations ]
-    29 [ label = public_HuddlePatientData ]
-    30 [ label = public_PmsAppointments ]
-    31 [ label = public_UnifiedAppointments ]
-    32 [ label = public_PmsTreatmentStatusHistory ]
-    33 [ label = public_PmsPatients ]
-    34 [ label = analyses_Calculus ]
-    35 [ label = public_UserEvents ]
-    36 [ label = public_PmsAppointmentToTreatments ]
-    37 [ label = analyses_Analyses ]
-    38 [ label = NonPhi_FolderHashToChartNum ]
-    39 [ label = public_PmsOperatories ]
-    40 [ label = analyses_Teeth ]
-    41 [ label = analyses_RawResponses ]
-    42 [ label = NonPhi_AnonymizedImageMetadata ]
-    43 [ label = public_ImagingPatientIdentifiers ]
-    44 [ label = analyses_AnalysesHistory ]
-    0 _> 40 [ ]
-    0 _> 37 [ ]
-    1 _> 5 [ ]
-    1 _> 33 [ ]
-    3 _> 5 [ ]
-    4 _> 3 [ ]
-    4 _> 11 [ ]
-    5 _> 28 [ ]
-    6 _> 3 [ ]
-    7 _> 4 [ ]
-    8 _> 40 [ ]
-    8 _> 37 [ ]
-    9 _> 5 [ ]
-    9 _> 33 [ ]
-    9 _> 3 [ ]
-    10 _> 4 [ ]
-    11 _> 5 [ ]
-    11 _> 3 [ ]
-    12 _> 28 [ ]
-    14 _> 5 [ ]
-    14 _> 28 [ ]
-    15 _> 37 [ ]
-    16 _> 3 [ ]
-    17 _> 11 [ ]
-    18 _> 40 [ ]
-    18 _> 37 [ ]
-    19 _> 5 [ ]
-    20 _> 5 [ ]
-    20 _> 33 [ ]
-    21 _> 4 [ ]
-    22 _> 5 [ ]
-    23 _> 40 [ ]
-    23 _> 37 [ ]
-    24 _> 40 [ ]
-    24 _> 37 [ ]
-    25 _> 40 [ ]
-    25 _> 37 [ ]
-    26 _> 5 [ ]
-    27 _> 5 [ ]
-    27 _> 2 [ ]
-    28 _> 5 [ ]
-    29 _> 33 [ ]
-    30 _> 14 [ ]
-    30 _> 5 [ ]
-    30 _> 33 [ ]
-    30 _> 39 [ ]
-    31 _> 5 [ ]
-    31 _> 33 [ ]
-    31 _> 30 [ ]
-    31 _> 28 [ ]
-    31 _> 3 [ ]
-    31 _> 11 [ ]
-    32 _> 1 [ ]
-    33 _> 5 [ ]
-    33 _> 28 [ ]
-    34 _> 40 [ ]
-    34 _> 37 [ ]
-    35 _> 11 [ ]
-    36 _> 1 [ ]
-    36 _> 30 [ ]
-    37 _> 41 [ ]
-    37 _> 4 [ ]
-    38 _> 5 [ ]
-    39 _> 5 [ ]
-    40 _> 37 [ ]
-    41 _> 4 [ ]
-    42 _> 13 [ ]
-    42 _> 4 [ ]
-    43 _> 5 [ ]
-    43 _> 3 [ ]
-    44 _> 4 [ ]
-}
