--- videahealth-dev/NonPhi.FolderHashToChartNum/data.csv
@@ -1,1 +1,0 @@
-id,practiceId,chartNum,extractedFolderPath,folderHash,createdAt,updatedAt,deletedAt
