--- videahealth-dev/Phi.PhiImageMetadata/data.csv
@@ -1,19 +1,0 @@
-id,otherPatientIDs,patientId,patientName,patientFirstName,patientLastName,sopClassUID,sopInstanceUID,seriesInstanceUID,institutionName,createdAt,updatedAt,deletedAt,patientBirthDate
-9885,,aud2gj3c15wa81343061ngt97wbfib,Mohammad Okeke,,,1.2.840.10008.5.1.4.1.1.1.3.1\0,2.25.0.395.10300006281994638647424651795504501890301789919450.1\0,2.25.0.395.11856666010687091256528207931234456861881557166668.1\0,Heartland,2023-12-08 19:44:23.608552+00,2023-12-08 19:44:23.608552+00,,
-9886,,aud2gj3c15wa81343061ngt97wbfib,Mohammad Okeke,,,1.2.840.10008.5.1.4.1.1.1.3.1\0,2.25.0.395.10300006281994638647424651795504501890301789919450.4\0,2.25.0.395.11856666010687091256528207931234456861881557166668.4\0,Heartland,2023-12-08 19:44:23.609719+00,2023-12-08 19:44:23.609719+00,,
-9887,,aud2gj3c15wa81343061ngt97wbfib,Mohammad Okeke,,,1.2.840.10008.5.1.4.1.1.1.3.1\0,2.25.0.395.10300006281994638647424651795504501890301789919450.2\0,2.25.0.395.11856666010687091256528207931234456861881557166668.2\0,Heartland,2023-12-08 19:44:23.612152+00,2023-12-08 19:44:23.612152+00,,
-9888,,aud2gj3c15wa81343061ngt97wbfib,Mohammad Okeke,,,1.2.840.10008.5.1.4.1.1.1.3.1\0,2.25.0.395.10300006281994638647424651795504501890301789919450.3\0,2.25.0.395.11856666010687091256528207931234456861881557166668.3\0,Heartland,2023-12-08 19:44:23.614683+00,2023-12-08 19:44:23.614683+00,,
-9889,,aud2gj3c15wa81343061ngt97wbfib,Mohammad Okeke,,,1.2.840.10008.5.1.4.1.1.1.3.1\0,2.25.0.395.10300006281994638647424651795504501890301789919450.0\0,2.25.0.395.11856666010687091256528207931234456861881557166668.0\0,Heartland,2023-12-08 19:44:23.617968+00,2023-12-08 19:44:23.617968+00,,
-9890,,aud2gj3c15wa81343061ngt97wbfib,Mohammad Okeke,,,1.2.840.10008.5.1.4.1.1.1.3.1\0,2.25.0.395.10300006281994638647424651795504501890301789919450.5\0,2.25.0.395.11856666010687091256528207931234456861881557166668.5\0,Heartland,2023-12-08 19:44:23.629452+00,2023-12-08 19:44:23.629452+00,,
-9891,,aud2gj3c15wa81343061ngt97wbfib,Mohammad Okeke,,,1.2.840.10008.5.1.4.1.1.1.3.1\0,2.25.0.395.10300006281994638647424651795504501890301789919450.11,2.25.0.395.11856666010687091256528207931234456861881557166668.11,Heartland,2023-12-08 19:44:23.98671+00,2023-12-08 19:44:23.98671+00,,
-9892,,aud2gj3c15wa81343061ngt97wbfib,Mohammad Okeke,,,1.2.840.10008.5.1.4.1.1.1.3.1\0,2.25.0.395.10300006281994638647424651795504501890301789919450.8\0,2.25.0.395.11856666010687091256528207931234456861881557166668.8\0,Heartland,2023-12-08 19:44:24.091705+00,2023-12-08 19:44:24.091705+00,,
-9893,,aud2gj3c15wa81343061ngt97wbfib,Mohammad Okeke,,,1.2.840.10008.5.1.4.1.1.1.3.1\0,2.25.0.395.10300006281994638647424651795504501890301789919450.14,2.25.0.395.11856666010687091256528207931234456861881557166668.14,Heartland,2023-12-08 19:44:28.294054+00,2023-12-08 19:44:28.294054+00,,
-9894,,aud2gj3c15wa81343061ngt97wbfib,Mohammad Okeke,,,1.2.840.10008.5.1.4.1.1.1.3.1\0,2.25.0.395.10300006281994638647424651795504501890301789919450.6\0,2.25.0.395.11856666010687091256528207931234456861881557166668.6\0,Heartland,2023-12-08 19:44:28.518771+00,2023-12-08 19:44:28.518771+00,,
-9895,,aud2gj3c15wa81343061ngt97wbfib,Mohammad Okeke,,,1.2.840.10008.5.1.4.1.1.1.3.1\0,2.25.0.395.10300006281994638647424651795504501890301789919450.16,2.25.0.395.11856666010687091256528207931234456861881557166668.16,Heartland,2023-12-08 19:44:28.647786+00,2023-12-08 19:44:28.647786+00,,
-9896,,aud2gj3c15wa81343061ngt97wbfib,Mohammad Okeke,,,1.2.840.10008.5.1.4.1.1.1.3.1\0,2.25.0.395.10300006281994638647424651795504501890301789919450.10,2.25.0.395.11856666010687091256528207931234456861881557166668.10,Heartland,2023-12-08 19:44:28.769333+00,2023-12-08 19:44:28.769333+00,,
-9897,,aud2gj3c15wa81343061ngt97wbfib,Mohammad Okeke,,,1.2.840.10008.5.1.4.1.1.1.3.1\0,2.25.0.395.10300006281994638647424651795504501890301789919450.9\0,2.25.0.395.11856666010687091256528207931234456861881557166668.9\0,Heartland,2023-12-08 19:44:28.774505+00,2023-12-08 19:44:28.774505+00,,
-9898,,aud2gj3c15wa81343061ngt97wbfib,Mohammad Okeke,,,1.2.840.10008.5.1.4.1.1.1.3.1\0,2.25.0.395.10300006281994638647424651795504501890301789919450.17,2.25.0.395.11856666010687091256528207931234456861881557166668.17,Heartland,2023-12-08 19:44:28.816078+00,2023-12-08 19:44:28.816078+00,,
-9899,,aud2gj3c15wa81343061ngt97wbfib,Mohammad Okeke,,,1.2.840.10008.5.1.4.1.1.1.3.1\0,2.25.0.395.10300006281994638647424651795504501890301789919450.12,2.25.0.395.11856666010687091256528207931234456861881557166668.12,Heartland,2023-12-08 19:44:29.074115+00,2023-12-08 19:44:29.074115+00,,
-9900,,aud2gj3c15wa81343061ngt97wbfib,Mohammad Okeke,,,1.2.840.10008.5.1.4.1.1.1.3.1\0,2.25.0.395.10300006281994638647424651795504501890301789919450.15,2.25.0.395.11856666010687091256528207931234456861881557166668.15,Heartland,2023-12-08 19:44:29.241494+00,2023-12-08 19:44:29.241494+00,,
-9901,,aud2gj3c15wa81343061ngt97wbfib,Mohammad Okeke,,,1.2.840.10008.5.1.4.1.1.1.3.1\0,2.25.0.395.10300006281994638647424651795504501890301789919450.13,2.25.0.395.11856666010687091256528207931234456861881557166668.13,Heartland,2023-12-08 19:44:32.114843+00,2023-12-08 19:44:32.114843+00,,
-9902,,aud2gj3c15wa81343061ngt97wbfib,Mohammad Okeke,,,1.2.840.10008.5.1.4.1.1.1.3.1\0,2.25.0.395.10300006281994638647424651795504501890301789919450.7\0,2.25.0.395.11856666010687091256528207931234456861881557166668.7\0,Heartland,2023-12-08 19:44:32.27129+00,2023-12-08 19:44:32.27129+00,,
