--- videahealth-dev/analyses.AnalysesHistory/data.csv
@@ -1,1 +1,0 @@
-id,createdAt,updatedAt,updatedBy,deletedAt,imageManifestId,data,note,type
