--- videahealth-dev/analyses.FmxSummary/data.csv
@@ -1,1 +1,0 @@
-id,createdAt,updatedAt,deletedAt,estimatedVisitId,toothId,dmf,hash
