--- videahealth-dev/public.AnalyzedImages/data.csv
@@ -1,1 +1,0 @@
-id,organizationId,practiceId,patientId,externalImageId,videaImageId,imageCreatedAt,externalExamId,imageMetaData,createdAt,updatedAt,deletedAt,createdBy,updatedBy,deletedBy
