--- videahealth-dev/public.EstimatedVisit/data.csv
@@ -1,2 +1,0 @@
-id,imageCreatedAt,imageCount,practiceId,patientId,createdAt,updatedAt,deletedAt
-7548,2023-12-08 19:43:07+00,18,6217dc7a04809c1afd525a78,5243,2023-12-08 19:44:35.745925+00,2023-12-08 19:44:56.62473+00,
