--- videahealth-dev/public.Practices/data.csv
@@ -1,2 +1,0 @@
-id,videaId,createdAt,updatedAt,deletedAt,name,location,organizationId,status,accountOwnerName,sensorTypes,pmsIntegrationType,pmsFileTypes,flags,inMongo,lastPatientTaskCalculationAt,timezone,userExperience,externalPracticeId,pmsShortName,isTest,onboardingDate,configuration,cloudwatchLogStream,country,language,dentrixSerialNumber
-665,6217dc7a04809c1afd525a78,2023-03-17 17:11:08.832139+00,2023-12-08 18:32:11.536654+00,,Videa Dev Default Practice,19 Kingston St,1168,Active,,[],DentiMax,"["".dcm""]","[{""key"": ""TOOTH_IDENTIFICATION_ENABLED"", ""value"": true}, {""key"": ""INSURANCE_CLAIM_SUMMARY_ENABLED"", ""value"": true}, {""key"": ""VCA_SUBTYPE_ENABLED"", ""value"": true}, {""key"": ""RBL_ENABLED"", ""value"": true}, {""key"": ""IMAGE_HISTORY_ENABLED"", ""value"": true}, {""key"": ""NEW_FMX_ALGORITHM_ENABLED"", ""value"": true}, {""key"": ""ROTATION_ENABLED"", ""value"": true}, {""key"": ""PA_CARIES_ENABLED"", ""value"": true}, {""key"": ""TOOTH_THIRDS_ENABLED"", ""value"": true}, {""key"": ""CARIES_DEPTH_ENABLED"", ""value"": true}, {""key"": ""PLANNING_MODE_DISABLED"", ""value"": true}, {""key"": ""NON_XRAY_VIEW_ENABLED"", ""value"": true}, {""key"": ""PRL_ENABLED"", ""value"": true}, {""key"": ""CALCULUS_ENABLED"", ""value"": true}, {""key"": ""PATIENT_VIEW_ENABLED"", ""value"": true}, {""key"": ""TOOTH_PARTS_SEGMENTATIONS_ENABLED"", ""value"": true}, {""key"": ""DDAI_METRICS_ENABLED"", ""value"": true}, {""key"": ""DMF_ENABLED"", ""value"": true}, {""key"": ""ONBOARDING_ENABLED"", ""value"": true}, {""key"": ""VIDEA_DENTAL_ASSIST_ENABLED"", ""value"": true}]",f,2023-08-02 20:54:38.959+00,,STANDALONE,,,f,,Standalone,,,,
